refactor : name 에 대한 test 를 Car 가 아닌 CarName 에서 진행
